[BugFix] Fix copy-paste error in moe_reduce.cu: FLOAT16 case incorrectly used BFLOAT16 template
The FLOAT16 switch case in MoeExpertReduceKernel was calling
MoeReduceKernel<paddle::DataType::BFLOAT16> instead of FLOAT16,
causing dtype mismatch errors on V100 (SM70) and any FP16 inference.

+++ custom_ops/gpu_ops/moe/moe_reduce.cu
@@ -90,17 +90,17 @@ paddle::Tensor MoeExpertReduceFunc(
                                                   &output);
       break;
     case paddle::DataType::FLOAT16:
-      MoeReduceKernel<paddle::DataType::BFLOAT16>(ffn_out,
-                                                  top_k_weight,
-                                                  permute_indices_per_token,
-                                                  top_k_indices,
-                                                  down_proj_bias,
-                                                  norm_topk_prob,
-                                                  routed_scaling_factor,
-                                                  num_rows,
-                                                  hidden_size,
-                                                  topk,
-                                                  &output);
+      MoeReduceKernel<paddle::DataType::FLOAT16>(ffn_out,
+                                                 top_k_weight,
+                                                 permute_indices_per_token,
+                                                 top_k_indices,
+                                                 down_proj_bias,
+                                                 norm_topk_prob,
+                                                 routed_scaling_factor,
+                                                 num_rows,
+                                                 hidden_size,
+                                                 topk,
+                                                 &output);
       break;
     default:
       PD_THROW("Unsupported data type for MoeDispatchKernel");
